+++ app/tests/integration/tox.ini
@@ -3,7 +3,6 @@ skipsdist = true
 
 [testenv:integration-test]
 passenv = *
-recreate = true
 commands =
     pip install -r ../requirements.txt
     pytest --verify-base-url --base-url https://nginx/ --driver Firefox . {posargs} -vvv
